SquareRootMatrix -> PSDMatrix
